swarch/requirements.txt
Flask==2.3.3 
Flask-SQLAlchemy==3.1.1 
mysqlclient==2.2.0 